Check colors available on your terminal
This is a script made by user l0b0 in which you can run from folder directory
that will allow you to display the number of colors as well as the actual
colors that are supported by your terminal.

I am adding this just a very minor side note for issue #16. If, by chance, you
are unsure if or how many colors your terminal support, here's an easy way to
find out.

While in the directory of the xterm-color-count, run the command
"./xterm-color-count.sh -v" and all the supported colors will be displayed.

Note: This is not supported on Linux.