Update color picking for user
This file is updated so that the user can choose in their testing
file to pick a corresponding color for their font. There are still
global variables in the file to choose background color of text,
highlight font color, and highlight font background color. Since
a user will usually just keep the background color of the text the
same, the user can just change the global variable for the font
background color while the font foreground color can be defined
by the user while they are defining their menu items.

Thanks to Kevin for the updated way to color each item separately!
My version allows the user to further color font background and
highlight font and background colors.