Filter menu by user input
The code has been updated to allow user input to filter
the menu. The filter can only occur for items on the current
menu. You can't filter an item that's within a submenu if you
are on the main menu and vise versa. This current version
doesn't support numbers in item names as those are used for
menu selection.