Filter user input including numbers
Update to filtering function to include numbers in title of
menu items. Previously, something like item1 could not be
filtered because 1 was used as a selector. Now you can filter
menu items with numbers in them. However, once you begin to
type in something to filter the menu by you can no longer
use the numbers to navigate the menu.